// lib/taxonomy/gics-entries.generated.ts
export type GicsEntry = {
  key: string;
  sectorId: string;
  sector: string;
  industryGroupId: string;
  industryGroup: string;
  industryId: string;
  industry: string;
  subIndustryId: string;
  subIndustry: string;
  subIndustryDescription: string;
};
// ...
export const GICS_ENTRY_KEYS = GICS_ENTRIES.map((e) => e.key) as [string, ...string[]];